Script adapted to know automatically the OS

+++ java_install.sh
@@ -1,30 +1,31 @@
 #!/bin/sh
-# Autor: Rodrigo Fioravante
+# Author: Rodrigo Fioravante
 # Email: kbsafioravante@gmail.com
 
-### Ubuntu ###
-# add-apt-repository ppa:webupd8team/java
-##############
+distro="$(lsb_release --id --short)"
 
-### Debian ###
+if [ "$distro" = "Debian" ]; then
+  distro_codename="$(lsb_release --id --short)"
+  if ["$distro_codename" = "stretch"]; then
+    apt remove gnupg
+    apt install --reinstall gnupg2
+    apt install dirmngr
 
-# Use this in Debian Stretch
-apt remove gnupg
-apt install --reinstall gnupg2
-apt install dirmngr
-#
+  echo -e \
+  "#Oracle Java 8 Repository
+  deb http://ppa.launchpad.net/webupd8team/java/ubuntu xenial main
+  deb-src http://ppa.launchpad.net/webupd8team/java/ubuntu xenial main"\
+  > /etc/apt/sources.list.d/java8.list
 
-echo -e \
-"#Oracle Java 8 Repository
-deb http://ppa.launchpad.net/webupd8team/java/ubuntu xenial main
-deb-src http://ppa.launchpad.net/webupd8team/java/ubuntu xenial main"\
-> /etc/apt/sources.list.d/java8.list
-
-apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys EEA14886
-##############
+  apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys EEA14886
+elif [ "$distro" = "Ubuntu" ]; then
+  add-apt-repository ppa:webupd8team/java
+elif [ "$distro" = "LinuxMint" ]; then
+  add-apt-repository ppa:webupd8team/java
+else
+  echo "Operating System doesn't known by script"
+fi
 
 apt update
-
 apt install oracle-java8-installer
-
 apt install oracle-java8-set-default
